requirements.txt
# Core framework dependencies
agno==2.0.9              # Agentic framework for agents and teams
pydantic>=2.11.0         # Data validation and type hints

# AI/ML model providers
openai>=1.109.0          # OpenAI API client (for GPT models)
google-generativeai>=0.8.5  # Google Gemini API client
google-genai

# Data processing and analysis
pandas>=2.3.0            # Data manipulation and analysis
numpy>=2.3.0             # Numerical computing
yfinance>=0.2.66         # Yahoo Finance data

# Environment and configuration
python-dotenv>=1.1.0     # Environment variable management

# Tracing and observability
langfuse>=3.5.0          # LLM tracing and observability

# Web search capabilities
exa-py>=1.15.6           # Exa search API (used by agno.tools.exa)